requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
Django==5.0
djangorestframework==3.14.0
dnspython==2.4.2
install==1.3.5
pycparser==2.21
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
